// src/app/shared/types/api-types/bidt-qualification-dto.interface.ts
export interface BidtQualificationDTO {
  qualificationCode?: string;
  qualificationTitle?: string;
  qualificationDegre?: string;
  qualificationDescription?: string;
  qualificationType?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  standard?: boolean;
  id?: number;
  averageHourRate?: number;
}